requirements.txt
flask>=2.0.2
pyyaml>=6.0
requests>=2.27.1
schedule>=1.1.0
